test-sample/requirements.txt
requests==2.25.0
urllib3==1.26.5
pillow==8.0.0
jinja2==2.11.0
flask==1.1.2
django==3.2.0
pyyaml==5.4.0
cryptography==3.4.0
